github-issues-to-project/pyproject.toml
[project]
name = "github-issues-to-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "typer>=0.24.1",
    "typer-config>=1.5.1",
]
